pyproject.toml
[project]
name = "assignemnt2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gtts>=2.5.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "manim>=0.19.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]
